+++ packages/api-client/src/views/Request/ResponseSection/ResponseEmpty.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { ScalarHotkey } from '@scalar/components'
 import type { Collection, Operation } from '@scalar/oas-utils/entities/spec'
 import type { Workspace } from '@scalar/oas-utils/entities/workspace'
 import { nextTick, onBeforeUnmount, onMounted } from 'vue'
@@ -7,7 +8,6 @@ import { useRoute, useRouter } from 'vue-router'
 import Computer from '@/assets/computer.ascii?raw'
 import EmptyState from '@/components/EmptyState.vue'
 import ScalarAsciiArt from '@/components/ScalarAsciiArt.vue'
-import ScalarHotkey from '@/components/ScalarHotkey.vue'
 import { useLayout } from '@/hooks'
 import type { HotKeyEvent } from '@/libs'
 import { useWorkspace } from '@/store'