+++ packages/components/src/components/ScalarHotkey/ScalarHotkey.e2e.ts
@@ -1,0 +1,5 @@
+import { testComponent } from '@test/testComponent'
+
+testComponent('ScalarHotkey', {
+  stories: ['Base', 'Multiple Modifiers'],
+})