+++ packages/components/src/components/ScalarTooltip/index.ts
@@ -1,4 +1,8 @@
 export { default as ScalarTooltip } from './ScalarTooltip.vue'
+export { default as ScalarHotkeyTooltip } from './ScalarHotkeyTooltip.vue'
 
-export type { ScalarTooltipPlacement, TooltipConfiguration } from './types'
+export type {
+  ScalarTooltipPlacement,
+  PublicTooltipConfiguration as TooltipConfiguration,
+} from './types'
 export { useTooltip } from './useTooltip'