+++ integrations/docker/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @scalarapi/docker-api-reference
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [97f199a]
+- Updated dependencies [0d502cb]
+- Updated dependencies [3de82ac]
+- Updated dependencies [58b9321]
+- Updated dependencies [2c74892]
+  - @scalar/api-reference@1.34.3
+
 ## 0.2.21
 
 ### Patch Changes