+++ integrations/nuxt/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/nuxt
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [b235392]
+- Updated dependencies [97f199a]
+- Updated dependencies [0d502cb]
+- Updated dependencies [3de82ac]
+- Updated dependencies [58b9321]
+- Updated dependencies [2c74892]
+- Updated dependencies [68e4631]
+- Updated dependencies [0afc40c]
+- Updated dependencies [128af48]
+- Updated dependencies [376739d]
+  - @scalar/api-client@2.5.23
+  - @scalar/api-reference@1.34.3
+  - @scalar/workspace-store@0.13.0
+  - @scalar/types@0.2.11
+  - @scalar/use-hooks@0.2.4
+
 ## 0.5.2
 
 ### Patch Changes