+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @scalar/api-client
 
+## 2.5.23
+
+### Patch Changes
+
+- b235392: feat: add de-select all button for scopes
+- 58b9321: feat: adds scalar code block copy component
+- 68e4631: fix(api-client): firefox input content field height
+- 376739d: fix: environment variable pill not displayed
+- Updated dependencies [e203e90]
+- Updated dependencies [3de82ac]
+- Updated dependencies [58b9321]
+- Updated dependencies [b1ceaab]
+- Updated dependencies [fbafeab]
+  - @scalar/openapi-parser@0.19.1
+  - @scalar/oas-utils@0.4.19
+  - @scalar/components@0.14.24
+  - @scalar/icons@0.4.7
+  - @scalar/import@0.4.17
+  - @scalar/postman-to-openapi@0.3.22
+  - @scalar/use-codemirror@0.12.25
+  - @scalar/draggable@0.2.0
+  - @scalar/helpers@0.0.7
+  - @scalar/object-utils@1.2.3
+  - @scalar/openapi-types@0.3.7
+  - @scalar/snippetz@0.4.4
+  - @scalar/themes@0.13.12
+  - @scalar/types@0.2.11
+  - @scalar/use-hooks@0.2.4
+  - @scalar/use-toasts@0.8.0
+  - @scalar/use-tooltip@1.1.0
+
 ## 2.5.22
 
 ### Patch Changes